Apparently, Chipmunk defaults to use doubles instead of floats...

+++ game/include/umbriel/physics/Space.hpp
@@ -12,6 +12,6 @@ namespace umbriel::physics
 
 		~Space() override = default;
 
-		void step(f32 delta);
+		void step(f64 delta);
 	};
 }
